# src/IO.Swagger/Controllers/DevolucionesApi.cs
/*
 * TPVV
 *
 * TPVV Ingeniería Web
 *
 * OpenAPI spec version: 1.0.0
 * Contact: lac56@alu.ua.es
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class DevolucionesApiController : ControllerBase
    { 
        /// <summary>
        /// Obtendremos los detalles de una devolucion en concreto
        /// </summary>
        /// <remarks>Obtendremos los detalles de una devolucion en concreto</remarks>
        /// <param name="token"></param>
        /// <param name="idPago"></param>
        /// <response code="200">Detalles sobre la transaccion</response>
        /// <response code="400">Esta respuesta significa que el servidor no pudo interpretar la solicitud dada una sintaxis inválida.</response>
        /// <response code="401">Sin autorización para realizar esta operación</response>
        /// <response code="404">No se encontró el recurso que se pidió</response>
        [HttpGet]
        [Route("/lac56-alu/TPVV/1.0.0/tpvv/detallesDevolucion")]
        [ValidateModelState]
        [SwaggerOperation("GetDetallesDevolucion")]
        [SwaggerResponse(statusCode: 200, type: typeof(Devoluciones), description: "Detalles sobre la transaccion")]
        [SwaggerResponse(statusCode: 400, type: typeof(InlineResponse400), description: "Esta respuesta significa que el servidor no pudo interpretar la solicitud dada una sintaxis inválida.")]
        [SwaggerResponse(statusCode: 401, type: typeof(InlineResponse401), description: "Sin autorización para realizar esta operación")]
        [SwaggerResponse(statusCode: 404, type: typeof(InlineResponse404), description: "No se encontró el recurso que se pidió")]
        public virtual IActionResult GetDetallesDevolucion([FromQuery][Required()]string token, [FromQuery][Required()]int? idPago)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Devoluciones));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(InlineResponse400));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(InlineResponse401));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default(InlineResponse404));
            string exampleJson = null;
            exampleJson = "{\n  \"detallesEstado\" : \"TARJETA DENEGADA\",\n  \"fecha\" : \"09-12-2022 09:55:23\",\n  \"total\" : 120.9,\n  \"estado\" : \"DEVOLUCIÓN\",\n  \"concepto\" : \"Compra memoria RAM\",\n  \"id\" : 1,\n  \"devolucionId\" : 22,\n  \"tarjeta\" : {\n    \"caducidad\" : \"09/2023\",\n    \"numero\" : \"984516545251\",\n    \"ccv\" : 123,\n    \"titularTarjeta\" : \"Luis Alfonso\"\n  },\n  \"referencia\" : \"515adfas54\"\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Devoluciones>(exampleJson)
                        : default(Devoluciones);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Se listan todas las devoluciones del usuario
        /// </summary>
        /// <remarks>Se obtienen una lista con todas las devoluciones que se han realizado por el usuario que hace la petición</remarks>
        /// <param name="token"></param>
        /// <response code="200">Lista de las devoluciones</response>
        /// <response code="401">Sin autorización para realizar esta operación</response>
        [HttpGet]
        [Route("/lac56-alu/TPVV/1.0.0/tpvv/listaDevoluciones")]
        [ValidateModelState]
        [SwaggerOperation("GetListaDevoluciones")]
        [SwaggerResponse(statusCode: 200, type: typeof(InlineResponse2001), description: "Lista de las devoluciones")]
        [SwaggerResponse(statusCode: 401, type: typeof(InlineResponse401), description: "Sin autorización para realizar esta operación")]
        public virtual IActionResult GetListaDevoluciones([FromQuery][Required()]string token)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(InlineResponse2001));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(InlineResponse401));
            string exampleJson = null;
            exampleJson = "{\n  \"data\" : [ {\n    \"detallesEstado\" : \"TARJETA DENEGADA\",\n    \"fecha\" : \"09-12-2022 09:55:23\",\n    \"total\" : 120.9,\n    \"estado\" : \"DEVOLUCIÓN\",\n    \"concepto\" : \"Compra memoria RAM\",\n    \"id\" : 1,\n    \"devolucionId\" : 22,\n    \"tarjeta\" : {\n      \"caducidad\" : \"09/2023\",\n      \"numero\" : \"984516545251\",\n      \"ccv\" : 123,\n      \"titularTarjeta\" : \"Luis Alfonso\"\n    },\n    \"referencia\" : \"515adfas54\"\n  }, {\n    \"detallesEstado\" : \"TARJETA DENEGADA\",\n    \"fecha\" : \"09-12-2022 09:55:23\",\n    \"total\" : 120.9,\n    \"estado\" : \"DEVOLUCIÓN\",\n    \"concepto\" : \"Compra memoria RAM\",\n    \"id\" : 1,\n    \"devolucionId\" : 22,\n    \"tarjeta\" : {\n      \"caducidad\" : \"09/2023\",\n      \"numero\" : \"984516545251\",\n      \"ccv\" : 123,\n      \"titularTarjeta\" : \"Luis Alfonso\"\n    },\n    \"referencia\" : \"515adfas54\"\n  } ]\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<InlineResponse2001>(exampleJson)
                        : default(InlineResponse2001);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Realizar una devolución sobre una de las compras realizadas
        /// </summary>
        /// <remarks>Realizar una devolución sobre una de las compras realizadas</remarks>
        /// <param name="token"></param>
        /// <param name="idPago"></param>
        /// <response code="201">Detalles sobre la transaccion</response>
        /// <response code="400">Esta respuesta significa que el servidor no pudo interpretar la solicitud dada una sintaxis inválida.</response>
        /// <response code="401">Sin autorización para realizar esta operación</response>
        /// <response code="404">No se encontró el recurso que se pidió</response>
        [HttpPost]
        [Route("/lac56-alu/TPVV/1.0.0/tpvv/realizarDevolucion")]
        [ValidateModelState]
        [SwaggerOperation("RealizarDevolucion")]
        [SwaggerResponse(statusCode: 201, type: typeof(Devoluciones), description: "Detalles sobre la transaccion")]
        [SwaggerResponse(statusCode: 400, type: typeof(InlineResponse400), description: "Esta respuesta significa que el servidor no pudo interpretar la solicitud dada una sintaxis inválida.")]
        [SwaggerResponse(statusCode: 401, type: typeof(InlineResponse401), description: "Sin autorización para realizar esta operación")]
        [SwaggerResponse(statusCode: 404, type: typeof(InlineResponse404), description: "No se encontró el recurso que se pidió")]
        public virtual IActionResult RealizarDevolucion([FromQuery][Required()]string token, [FromQuery][Required()]int? idPago)
        { 
            //TODO: Uncomment the next line to return response 201 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(201, default(Devoluciones));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(InlineResponse400));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(InlineResponse401));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default(InlineResponse404));
            string exampleJson = null;
            exampleJson = "{\n  \"detallesEstado\" : \"TARJETA DENEGADA\",\n  \"fecha\" : \"09-12-2022 09:55:23\",\n  \"total\" : 120.9,\n  \"estado\" : \"DEVOLUCIÓN\",\n  \"concepto\" : \"Compra memoria RAM\",\n  \"id\" : 1,\n  \"devolucionId\" : 22,\n  \"tarjeta\" : {\n    \"caducidad\" : \"09/2023\",\n    \"numero\" : \"984516545251\",\n    \"ccv\" : 123,\n    \"titularTarjeta\" : \"Luis Alfonso\"\n  },\n  \"referencia\" : \"515adfas54\"\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Devoluciones>(exampleJson)
                        : default(Devoluciones);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}