# src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "TPVV",
    "description" : "TPVV Ingeniería Web",
    "contact" : {
      "email" : "lac56@alu.ua.es"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/lac56-alu/TPVV/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  }, {
    "url" : "http://localhost:2000",
    "description" : "Desarrollo servidor basado en TPVV IW"
  }, {
    "url" : "",
    "description" : "Servidor"
  } ],
...
  "paths" : {
...
    "/tpvv/realizarDevolucion" : {
      "post" : {
        "tags" : [ "Devoluciones" ],
        "summary" : "Realizar una devolución sobre una de las compras realizadas",
        "description" : "Realizar una devolución sobre una de las compras realizadas",
        "operationId" : "realizarDevolucion",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "idPago",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Detalles sobre la transaccion",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Devoluciones"
                }
              }
            }
          },
          "400" : {
            "description" : "Esta respuesta significa que el servidor no pudo interpretar la solicitud dada una sintaxis inválida.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_400"
                }
              }
            }
          },
          "401" : {
            "description" : "Sin autorización para realizar esta operación",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_401"
                }
              }
            }
          },
          "404" : {
            "description" : "No se encontró el recurso que se pidió",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_404"
                }
              }
            }
          }
        }
      }
    },
    "/tpvv/listaDevoluciones" : {
      "get" : {
        "tags" : [ "Devoluciones" ],
        "summary" : "Se listan todas las devoluciones del usuario",
        "description" : "Se obtienen una lista con todas las devoluciones que se han realizado por el usuario que hace la petición",
        "operationId" : "getListaDevoluciones",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Lista de las devoluciones",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200_1"
                }
              }
            }
          },
          "401" : {
            "description" : "Sin autorización para realizar esta operación",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_401"
                }
              }
            }
          }
        }
      }
    },
    "/tpvv/detallesDevolucion" : {
      "get" : {
        "tags" : [ "Devoluciones" ],
        "summary" : "Obtendremos los detalles de una devolucion en concreto",
        "description" : "Obtendremos los detalles de una devolucion en concreto",
        "operationId" : "getDetallesDevolucion",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "idPago",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Detalles sobre la transaccion",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Devoluciones"
                }
              }
            }
          },
          "400" : {
            "description" : "Esta respuesta significa que el servidor no pudo interpretar la solicitud dada una sintaxis inválida.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_400"
                }
              }
            }
          },
          "401" : {
            "description" : "Sin autorización para realizar esta operación",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_401"
                }
              }
            }
          },
          "404" : {
            "description" : "No se encontró el recurso que se pidió",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_404"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Tarjeta" : {
        "type" : "object",
        "properties" : {
          "titularTarjeta" : {
            "type" : "string",
            "example" : "Luis Alfonso"
          },
          "numero" : {
            "type" : "string",
            "description" : "Número de la tarjeta de crédito",
            "example" : "984516545251"
          },
          "ccv" : {
            "type" : "number",
            "description" : "Código de seguridad",
            "example" : 123
          },
          "caducidad" : {
            "required" : [ "caducidad", "ccv", "numero", "titularTarjeta" ],
            "type" : "string",
            "description" : "Fecha de caducidad de la tarjeta",
            "format" : "MM/YYYY",
            "example" : "09/2023"
          }
        },
        "example" : {
          "caducidad" : "09/2023",
          "numero" : "984516545251",
          "ccv" : 123,
          "titularTarjeta" : "Luis Alfonso"
        }
      },
...
      "Devoluciones" : {
        "required" : [ "referencia", "tarjeta", "total" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "example" : 1
          },
          "devolucionId" : {
            "type" : "array",
            "description" : "Id de la devolución, si hubiera (en caso contrario -> null)",
            "example" : 22,
            "items" : {
              "type" : "integer"
            }
          },
          "total" : {
            "type" : "number",
            "description" : "Total de la compra",
            "example" : 120.9
          },
          "concepto" : {
            "type" : "string",
            "description" : "Se podrá añadir opcionalmente a la hora de realizar el pago un concepto sobre la compra realizada",
            "example" : "Compra memoria RAM"
          },
          "referencia" : {
            "type" : "string",
            "description" : "Es la referencia que vincula el pago con el recibo creado por el usuario que realiza la compra",
            "example" : "515adfas54"
          },
          "fecha" : {
            "type" : "string",
            "description" : "Fecha del momento en el que se realiza la compra",
            "format" : "DD-MM-YYYY hh:mm:ss",
            "example" : "09-12-2022 09:55:23"
          },
          "estado" : {
            "type" : "string",
            "description" : "Estado de como se encuentra el pago",
            "example" : "DEVOLUCIÓN",
            "enum" : [ "EN_ESPERA", "ACEPTADO", "DENEGADO, DEVOLUCIÓN" ]
          },
          "detallesEstado" : {
            "type" : "string",
            "description" : "Motivo del error del pago",
            "example" : "TARJETA DENEGADA"
          },
          "tarjeta" : {
            "$ref" : "#/components/schemas/Tarjeta"
          }
        },
        "example" : {
          "detallesEstado" : "TARJETA DENEGADA",
          "fecha" : "09-12-2022 09:55:23",
          "total" : 120.9,
          "estado" : "DEVOLUCIÓN",
          "concepto" : "Compra memoria RAM",
          "id" : 1,
          "devolucionId" : 22,
          "tarjeta" : {
            "caducidad" : "09/2023",
            "numero" : "984516545251",
            "ccv" : 123,
            "titularTarjeta" : "Luis Alfonso"
          },
          "referencia" : "515adfas54"
        }
      },
      "inline_response_400" : {
        "type" : "object",
        "properties" : {
          "typeError" : {
            "type" : "string",
            "example" : "ERROR_PARAMETERS"
          },
          "messageError" : {
            "type" : "string",
            "example" : "Los parámetros introducidos son erroneos."
          }
        }
      },
...
      "inline_response_401" : {
        "type" : "object",
        "properties" : {
          "errorMessage" : {
            "type" : "string",
            "example" : "ERROR_TOKEN"
          }
        }
      },
...
      "inline_response_404" : {
        "type" : "object",
        "properties" : {
          "errorMessage" : {
            "type" : "string",
            "example" : "ERROR_RECURSO_NO_ENCONTRADO"
          }
        }
      },
      "inline_response_200_1" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Devoluciones"
            }
          }
        },
        "example" : {
          "data" : [ {
            "detallesEstado" : "TARJETA DENEGADA",
            "fecha" : "09-12-2022 09:55:23",
            "total" : 120.9,
            "estado" : "DEVOLUCIÓN",
            "concepto" : "Compra memoria RAM",
            "id" : 1,
            "devolucionId" : 22,
            "tarjeta" : {
              "caducidad" : "09/2023",
              "numero" : "984516545251",
              "ccv" : 123,
              "titularTarjeta" : "Luis Alfonso"
            },
            "referencia" : "515adfas54"
          }, {
            "detallesEstado" : "TARJETA DENEGADA",
            "fecha" : "09-12-2022 09:55:23",
            "total" : 120.9,
            "estado" : "DEVOLUCIÓN",
            "concepto" : "Compra memoria RAM",
            "id" : 1,
            "devolucionId" : 22,
            "tarjeta" : {
              "caducidad" : "09/2023",
              "numero" : "984516545251",
              "ccv" : 123,
              "titularTarjeta" : "Luis Alfonso"
            },
            "referencia" : "515adfas54"
          } ]
        }
      },
...
    },
...
  }
}